remove IRotateParticleSystem.cs
