--- Assets/VRM10_Samples/ClothSample/RotateParticle/Runtime/IRotateParticleSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5c151a18a42e3714fbf6f52b128ce3e5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
