--- Assets/VRM10_Samples/ClothSample/RotateParticle/Runtime/RotateParticleJobRuntime.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2349993b3f7dcfb48b30630bbc55bd8b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
