--- Assets/VRM10_Samples/ClothSample/RotateParticle/Runtime/RotateParticleSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 67a32b7c5e169f64daa6a8717d3de4f1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
